Functions Continued Lab
In this lab I created three functions. The first function is called saturdayFun that takes in a parameter of activity with a default set to roller-skate. within that functions code block it returns a string and usese interpolation to pass in the parameter. The second function is called mondayWork and takes in one parameter with a default of go to the office. within that code block we again use interpolation to pass in the parameter. The third function is called wrapAdjective that takes in one parameter that has a default of a symbol and within that code block returns another function that returns a string that uses interpolation to pass in the parameters.